+++ doc/.vitepress/config.mjs
@@ -1,26 +1,43 @@
 import { defineConfig } from "vitepress";
 import { execSync } from "node:child_process";
 
-const isMain = process.env.IS_RELEASE !== "true";
-const version = execSync("git tag --list --sort=-v:refname")
-  .toString()
-  .split("\n")[0]
-  .trim();
+const inProd = process.env.NODE_ENV === "production";
 
-const siteUrl = "https://codecompanion.github.io";
+let version = "Main";
+if (inProd) {
+  try {
+    version = execSync("git describe --tags --abbrev=0", {
+      encoding: "utf-8",
+    }).trim();
+  } catch (error) {
+    console.warn("Failed to get git version, using default.");
+  }
+}
 
-const title = isMain ? "Main" : version;
-const otherTitle = isMain ? version : "Main";
+const baseHeaders = [];
+const umamiScript = [
+  "script",
+  {
+    defer: "true",
+    src: "https://cloud.umami.is/script.js",
+    "data-website-id": "6fb6c149-1aba-4531-b613-7fc54d42191a",
+  },
+];
+const headers = inProd ? [baseHeaders, umamiScript] : baseHeaders;
+
+const siteUrl = "https://codecompanion.olimorris.dev";
 
 // https://vitepress.dev/reference/site-config
 export default defineConfig({
-  title: "CodeCompanion.nvim",
+  title: "CodeCompanion",
   description: "AI-powered coding, seamlessly in Neovim",
+  head: headers,
   sitemap: { hostname: siteUrl },
   themeConfig: {
+    logo: "https://github.com/user-attachments/assets/825fc040-9bc8-4743-be2a-71e257f8a7be",
     nav: [
       {
-        text: version,
+        text: `${version}`,
         items: [
           {
             text: "Changelog",
@@ -36,16 +53,17 @@ export default defineConfig({
 
     sidebar: [
       { text: "Introduction", link: "/" },
-      { text: "Installation", link: "installation" },
-      { text: "Getting Started", link: "getting-started" },
+      { text: "Installation", link: "/installation" },
+      { text: "Getting Started", link: "/getting-started" },
       {
         text: "Configuration",
-        collapsed: false,
+        collapsed: true,
         items: [
+          { text: "Introduction", link: "/configuration/introduction" },
+          { text: "Action Palette", link: "/configuration/action-palette" },
           { text: "Adapters", link: "/configuration/adapters" },
           { text: "Chat Buffer", link: "/configuration/chat-buffer" },
           { text: "Inline Assistant", link: "/configuration/inline-assistant" },
-          { text: "Action Palette", link: "/configuration/action-palette" },
           { text: "Prompt Library", link: "/configuration/prompt-library" },
           { text: "System Prompt", link: "/configuration/system-prompt" },
           { text: "Others", link: "/configuration/others" },
@@ -53,27 +71,37 @@ export default defineConfig({
       },
       {
         text: "Usage",
-        collapsed: true,
+        collapsed: false,
         items: [
-          { text: "General", link: "/usage/general" },
-          { text: "Chat Buffer", link: "/usage/chat-buffer" },
-          { text: "Inline Assistant", link: "/usage/inline-assistant" },
-          { text: "Commands", link: "/usage/commands" },
+          { text: "Introduction", link: "/usage/introduction" },
           { text: "Action Palette", link: "/usage/action-palette" },
-          { text: "Adapters", link: "/usage/adapters" },
-          { text: "Agents/Tools", link: "/usage/agents" },
+          {
+            text: "Chat Buffer",
+            link: "/usage/chat-buffer/",
+            collapsed: true,
+            items: [
+              { text: "Agents/Tools", link: "/usage/chat-buffer/agents" },
+              {
+                text: "Slash Commands",
+                link: "/usage/chat-buffer/slash-commands",
+              },
+              { text: "Variables", link: "/usage/chat-buffer/variables" },
+            ],
+          },
           { text: "Events", link: "/usage/events" },
+          { text: "Inline Assistant", link: "/usage/inline-assistant" },
           { text: "Workflows", link: "/usage/workflows" },
-          { text: "Miscellaneous", link: "/usage/misc" },
+          { text: "Others", link: "/usage/others" },
         ],
       },
       {
         text: "Extending the Plugin",
-        collapsed: true,
+        collapsed: false,
         items: [
           { text: "Creating Adapters", link: "/extending/adapters" },
           { text: "Creating Prompts", link: "/extending/prompts" },
           { text: "Creating Tools", link: "/extending/tools" },
+          { text: "Creating Workspaces", link: "/extending/workspace" },
         ],
       },
     ],
@@ -95,5 +123,7 @@ export default defineConfig({
         link: "https://github.com/olimorris/codecompanion.nvim",
       },
     ],
+
+    search: { provider: "local" },
   },
 });