+++ lua/codecompanion/config.lua
@@ -177,6 +177,13 @@ Points to note:
             contains_code = false,
           },
         },
+        ["workspace"] = {
+          callback = "strategies.chat.slash_commands.workspace",
+          description = "Load a workspace file",
+          opts = {
+            contains_code = true,
+          },
+        },
       },
       keymaps = {
         options = {
@@ -261,11 +268,19 @@ Points to note:
           callback = "keymaps.pin_reference",
           description = "Pin Reference",
         },
+        watch = {
+          modes = {
+            n = "gw",
+          },
+          index = 10,
+          callback = "keymaps.toggle_watch",
+          description = "Watch Buffer",
+        },
         next_chat = {
           modes = {
             n = "}",
           },
-          index = 10,
+          index = 11,
           callback = "keymaps.next_chat",
           description = "Next Chat",
         },
@@ -273,7 +288,7 @@ Points to note:
           modes = {
             n = "{",
           },
-          index = 11,
+          index = 12,
           callback = "keymaps.previous_chat",
           description = "Previous Chat",
         },
@@ -281,7 +296,7 @@ Points to note:
           modes = {
             n = "]]",
           },
-          index = 12,
+          index = 13,
           callback = "keymaps.next_header",
           description = "Next Header",
         },
@@ -289,7 +304,7 @@ Points to note:
           modes = {
             n = "[[",
           },
-          index = 13,
+          index = 14,
           callback = "keymaps.previous_header",
           description = "Previous Header",
         },
@@ -297,7 +312,7 @@ Points to note:
           modes = {
             n = "ga",
           },
-          index = 14,
+          index = 15,
           callback = "keymaps.change_adapter",
           description = "Change adapter",
         },
@@ -305,7 +320,7 @@ Points to note:
           modes = {
             n = "gf",
           },
-          index = 14,
+          index = 15,
           callback = "keymaps.fold_code",
           description = "Fold code",
         },
@@ -313,7 +328,7 @@ Points to note:
           modes = {
             n = "gd",
           },
-          index = 15,
+          index = 16,
           callback = "keymaps.debug",
           description = "View debug info",
         },
@@ -321,7 +336,7 @@ Points to note:
           modes = {
             n = "gs",
           },
-          index = 16,
+          index = 17,
           callback = "keymaps.toggle_system_prompt",
           description = "Toggle the system prompt",
         },
@@ -817,6 +832,7 @@ This is the code, for context:
     chat = {
       icons = {
         pinned_buffer = " ",
+        watched_buffer = "👀 ",
       },
       window = {
         layout = "vertical", -- float|vertical|horizontal|buffer
@@ -878,10 +894,12 @@ This is the code, for context:
     ---@return boolean
     send_code = true,
 
-    -- This is the default prompt which is sent with every request in the chat
-    -- strategy. It is primarily based on the GitHub Copilot Chat's prompt
-    -- but with some modifications. You can choose to remove this via
-    -- your own config but note that LLM results may not be as good
+    ---This is the default prompt which is sent with every request in the chat
+    ---strategy. It is primarily based on the GitHub Copilot Chat's prompt
+    ---but with some modifications. You can choose to remove this via
+    ---your own config but note that LLM results may not be as good
+    ---@param opts table
+    ---@return string
     system_prompt = function(opts)
       local language = opts.language or "English"
       return string.format(